User Interface: Button constructor now takes dictionary
There were too many properties to not label them, now a Button must
be constructed like so:

    new Button do
      name: \name
      tag: 0
      size: 20

Also fixed the slight offset the typesetting added to the label that
made it appear to sit slightly lower than central.

Now that parameters come from a dictionary, must change `enabled` to
`@enabled` as the former is no longer available.