Change pan behaviour
To pan you now actually have to grab, there's a test to detect if
fingers are visible (no fingers visible on a fist).

This means you can move your hand in safely and then grab, and let go
when you want to stop.  This doesn't provide protection from panning by
someones head.

Also moved the zoom tolerance constant to where it should be.