Interact Nodes: BUGFIX: Zooming.
The wire being in a separate layer (specifically, having a different
center) to the view layer caused problems in scaling, so the wires are
now all inserted into a group in the view-layer, and everything else is
added above them.

Additionally, adding elements after scaling occurs also causes issues,
so the system now keeps track of the total scale factor and applies it
to all incoming child items.